PM-495 Unify markdown table converter for Jira and Github
